Added Flexbox
This site is currently under construction. With more content added, the site will be finalized at its basic level.

The color scheme change now works, allowing a user to change the color scheme of the entire page. 

Tooltips and a flexbox are now added to Color Scheme Changes and other minor changes.